// src/components/slider/types.tsx
export enum Classnames {
  ANIMATED = "animated animated-out fadeOut",
  ROOT = "slider",
  WRAPPER = "slider__wrapper",
  STAGE = "slider__stage",
  STAGE_ITEM = "slider__stage-item",
  DOTS = "slider__dots",
  DOTS_ITEM = "slider__dots-item",
  PLAY_BTN = "slider__play-btn",
  PLAY_BTN_ITEM = "slider__play-btn-item",
  PLAY_BTN_WRAPPER = "slider__play-btn-wrapper",
  SLIDE_INFO = "slider__slide-info",
  SLIDE_INFO_ITEM = "slider__slide-info-item",
  BUTTON_PREV = "slider__prev-btn",
  BUTTON_PREV_WRAPPER = "slider__prev-btn-wrapper",
  BUTTON_PREV_ITEM = "slider__prev-btn-item",
  BUTTON_NEXT = "slider__next-btn",
  BUTTON_NEXT_WRAPPER = "slider__next-btn-wrapper",
  BUTTON_NEXT_ITEM = "slider__next-btn-item",
}

export enum Modifiers {
  ACTIVE = "__active",
  INACTIVE = "__inactive",
  CLONED = "__cloned",
  CUSTOM = "__custom",
  PAUSE = "__pause",
  SEPARATOR = "__separator",
  SSR = "__ssr",
  TARGET = "__target",
}